move tracer implementation to its own file
